+++ src/components/forms/CheckboxGroup/styles.ts
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react'
-import { Theme } from '../../contexts/ThemeContext'
+import { Theme } from '../../../contexts/ThemeContext'
 
 export const listStyles = css`
   padding: 0;