+++ src/components/forms/DatePickerField/CalendarContainer/styles.ts
@@ -2,7 +2,7 @@ import { css } from '@emotion/react'
 
 import { N100_COLOR } from '../../../../colors'
 import { SANSSERIF_FONTSET } from '../../../../fonts'
-import { Theme } from '../../../contexts/ThemeContext'
+import { Theme } from '../../../../contexts/ThemeContext'
 
 const buttonWidth = 40