+++ src/components/forms/FormErrorMessages/styles.ts
@@ -2,7 +2,7 @@ import { css } from '@emotion/react'
 
 import { N300_COLOR } from '../../../colors'
 import { SANSSERIF_FONTSET } from '../../../fonts'
-import { Theme } from '../../contexts/ThemeContext'
+import { Theme } from '../../../contexts/ThemeContext'
 
 export const errorStyle = css`
   margin-top: 10px;