+++ src/components/forms/MarkdownField/styles.ts
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react'
 import { N100_COLOR, N300_COLOR } from '../../../colors'
-import { Theme } from '../../contexts/ThemeContext'
+import { Theme } from '../../../contexts/ThemeContext'
 
 export const textStyle = css`
   margin: 0;