+++ src/components/layout/Tabs/styles.ts
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react'
 
-import { Theme } from '../../contexts/ThemeContext'
+import { Theme } from '../../../contexts/ThemeContext'
 
 export const defaultTabSyles = (theme: Theme) => css`
   padding: 5px;