+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useMergedFocusHandlers } from './useMergedFocusHandlers'
 export { useSamenessNumberState } from './useSamenessNumberState'
 export { useTheme } from './useTheme'
+export { useThemeMutation } from './useThemeMutation'