+++ src/hooks/useThemeMutation.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { ThemeMutationContext } from '../contexts/ThemeContext'
+
+export const useThemeMutation = () => useContext(ThemeMutationContext)